move Direction to types
more work to clean up things for futuer updates
